requirements.txt
flask
google-cloud-translate
polib
